backend/zapsync_project/zapsync_app/users/serializers.py
from rest_framework import serializers
from .models import User

class RegisterUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'  # Include all fields in the User model
