backend/requirements.txt
django
djangorestframework
numpy
pandas
scikit-learn
matplotlib
joblib
seaborn
pydantic
requests
imblearn
mysqlclient 
social-auth-app-django 
djangorestframework-simplejwt
django-cors-headers
django-allauth
bcrypt